added a watch mode

+++ assembly/index.ts
@@ -3,5 +3,5 @@ function double(a:f32): f32 {
 }
 
 export function add(a:f32, b: f32): f32 {
-    return double(a) + b;
+    return double(a) + double(b);
 }
